+++ tests/unit/plugins/test_plugin_compliance.py
@@ -20,6 +20,7 @@ from poriscope.utils.MetaFilter import MetaFilter
 from poriscope.utils.MetaModel import MetaModel
 from poriscope.utils.MetaReader import MetaReader
 from poriscope.utils.MetaSubsetTabController import MetaSubsetTabController
+from poriscope.utils.MetaSubsetTabModel import MetaSubsetTabModel
 from poriscope.utils.MetaSubsetTabView import MetaSubsetTabView
 from poriscope.utils.MetaView import MetaView
 from poriscope.utils.MetaWriter import MetaWriter
@@ -74,6 +75,7 @@ INTERMEDIATE_BASES: Set[Type] = {
     MetaEventTabController,
     MetaEventTabView,
     MetaSubsetTabController,
+    MetaSubsetTabModel,
     MetaSubsetTabView,
 }